requirements.txt
PyExifTool==0.5.6
python-dateutil>=2.8.2
olefile>=0.46
hachoir>=3.3.0
textract>=1.6.5
